+++ src/tests/golden/testdata/underscore/golden.txt
@@ -34,27 +34,24 @@ MAYBE   arraySyntax: Use of old array syntax (use short form instead) at testdat
 INFO    unused: Variable is_sorted is unused (use $_ to ignore this inspection) at testdata/underscore/underscore.php:356
     list($collection, $is_sorted, $iterator) = self::_wrapArgs(func_get_args(), 3);
                       ^^^^^^^^^^
-MAYBE   arraySyntax: Use of old array syntax (use short form instead) at testdata/underscore/underscore.php:448
-        list($start, $stop, $step) = array(0, $args[0], 1);
-                                     ^^^^^^^^^^^^^^^^^^^^^
 MAYBE   arrayAccess: Array access to non-array type \__|mixed at testdata/underscore/underscore.php:448
         list($start, $stop, $step) = array(0, $args[0], 1);
                                               ^^^^^
-MAYBE   arraySyntax: Use of old array syntax (use short form instead) at testdata/underscore/underscore.php:451
-        list($start, $stop, $step) = array($args[0], $args[1], 1);
-                                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+MAYBE   arraySyntax: Use of old array syntax (use short form instead) at testdata/underscore/underscore.php:448
+        list($start, $stop, $step) = array(0, $args[0], 1);
+                                     ^^^^^^^^^^^^^^^^^^^^^
 MAYBE   arrayAccess: Array access to non-array type \__|mixed at testdata/underscore/underscore.php:451
         list($start, $stop, $step) = array($args[0], $args[1], 1);
                                            ^^^^^
 MAYBE   arrayAccess: Array access to non-array type \__|mixed at testdata/underscore/underscore.php:451
         list($start, $stop, $step) = array($args[0], $args[1], 1);
                                                      ^^^^^
+MAYBE   arraySyntax: Use of old array syntax (use short form instead) at testdata/underscore/underscore.php:451
+        list($start, $stop, $step) = array($args[0], $args[1], 1);
+                                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 MAYBE   arraySyntax: Use of old array syntax (use short form instead) at testdata/underscore/underscore.php:452
         if($stop < $start) return self::_wrap(array());
                                               ^^^^^^^
-MAYBE   arraySyntax: Use of old array syntax (use short form instead) at testdata/underscore/underscore.php:455
-        list($start, $stop, $step) = array($args[0], $args[1], $args[2]);
-                                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 MAYBE   arrayAccess: Array access to non-array type \__|mixed at testdata/underscore/underscore.php:455
         list($start, $stop, $step) = array($args[0], $args[1], $args[2]);
                                            ^^^^^
@@ -64,6 +61,9 @@ MAYBE   arrayAccess: Array access to non-array type \__|mixed at testdata/unders
 MAYBE   arrayAccess: Array access to non-array type \__|mixed at testdata/underscore/underscore.php:455
         list($start, $stop, $step) = array($args[0], $args[1], $args[2]);
                                                                ^^^^^
+MAYBE   arraySyntax: Use of old array syntax (use short form instead) at testdata/underscore/underscore.php:455
+        list($start, $stop, $step) = array($args[0], $args[1], $args[2]);
+                                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 MAYBE   arraySyntax: Use of old array syntax (use short form instead) at testdata/underscore/underscore.php:456
         if($step > 0 && $step > $stop) return self::_wrap(array($start));
                                                           ^^^^^^^^^^^^^
